@@ -4,12 +4,14 @@
 # See the NOTICE for more information.
 
 import os
+import re
 import sys
 
 import pytest
 
 from gunicorn import config
 from gunicorn.app.base import Application
+from gunicorn.app.wsgiapp import WSGIApplication
 from gunicorn.errors import ConfigError
 from gunicorn.workers.sync import SyncWorker
 from gunicorn import glogging
@@ -24,6 +26,8 @@ def cfg_file():
     return os.path.join(dirname, "config", "test_cfg.py")
 def alt_cfg_file():
     return os.path.join(dirname, "config", "test_cfg_alt.py")
+def cfg_file_with_wsgi_app():
+    return os.path.join(dirname, "config", "test_cfg_with_wsgi_app.py")
 def paster_ini():
     return os.path.join(dirname, "..", "examples", "frameworks", "pylonstest", "nose.ini")
 
@@ -42,10 +46,18 @@ class AltArgs(object):
 
 class NoConfigApp(Application):
     def __init__(self):
-        super(NoConfigApp, self).__init__("no_usage", prog="gunicorn_test")
+        super().__init__("no_usage", prog="gunicorn_test")
 
     def init(self, parser, opts, args):
         pass
+
+    def load(self):
+        pass
+
+
+class WSGIApp(WSGIApplication):
+    def __init__(self):
+        super().__init__("no_usage", prog="gunicorn_test")
 
     def load(self):
         pass
@@ -352,11 +364,40 @@ def test_invalid_enviroment_variables_config(monkeypatch, capsys):
         _, err = capsys.readouterr()
         assert  "error: unrecognized arguments: --foo" in err
 
+
 def test_cli_overrides_enviroment_variables_module(monkeypatch):
     monkeypatch.setenv("GUNICORN_CMD_ARGS", "--workers=4")
     with AltArgs(["prog_name", "-c", cfg_file(), "--workers", "3"]):
         app = NoConfigApp()
     assert app.cfg.workers == 3
+
+
+@pytest.mark.parametrize("options, expected", [
+    (["app:app"], 'app:app'),
+    (["-c", cfg_file(), "app:app"], 'app:app'),
+    (["-c", cfg_file_with_wsgi_app(), "app:app"], 'app:app'),
+    (["-c", cfg_file_with_wsgi_app()], 'app1:app1'),
+])
+def test_wsgi_app_config(options, expected):
+    cmdline = ["prog_name"]
+    cmdline.extend(options)
+    with AltArgs(cmdline):
+        app = WSGIApp()
+    assert app.app_uri == expected
+
+
+@pytest.mark.parametrize("options", [
+    ([]),
+    (["-c", cfg_file()]),
+])
+def test_non_wsgi_app(options, capsys):
+    cmdline = ["prog_name"]
+    cmdline.extend(options)
+    with AltArgs(cmdline):
+        with pytest.raises(SystemExit):
+            WSGIApp()
+        _, err = capsys.readouterr()
+        assert  "Error: No application module specified." in err
 
 
 @pytest.mark.parametrize("options, expected", [
@@ -429,3 +470,51 @@ def _test_ssl_version(options, expected):
     with AltArgs(cmdline):
         app = NoConfigApp()
     assert app.cfg.ssl_version == expected
+
+
+def test_bind_fd():
+    with AltArgs(["prog_name", "-b", "fd://42"]):
+        app = NoConfigApp()
+    assert app.cfg.bind == ["fd://42"]
+
+
+def test_repr():
+    c = config.Config()
+    c.set("workers", 5)
+
+    assert "with value 5" in repr(c.settings['workers'])
+
+
+def test_str():
+    c = config.Config()
+    o = str(c)
+
+    # match the first few lines, some different types, but don't go OTT
+    # to avoid needless test fails with changes
+    OUTPUT_MATCH = {
+        'access_log_format': '%(h)s %(l)s %(u)s %(t)s "%(r)s" %(s)s %(b)s "%(f)s" "%(a)s"',
+        'accesslog': 'None',
+        'backlog': '2048',
+        'bind': "['127.0.0.1:8000']",
+        'capture_output': 'False',
+        'child_exit': '<ChildExit.child_exit()>',
+    }
+    for i, line in enumerate(o.splitlines()):
+        m = re.match(r'^(\w+)\s+= ', line)
+        assert m, "Line {} didn't match expected format: {!r}".format(i, line)
+
+        key = m.group(1)
+        try:
+            s = OUTPUT_MATCH.pop(key)
+        except KeyError:
+            continue
+
+        line_re = r'^{}\s+= {}$'.format(key, re.escape(s))
+        assert re.match(line_re, line), '{!r} != {!r}'.format(line_re, line)
+
+        if not OUTPUT_MATCH:
+            break
+    else:
+        assert False, 'missing expected setting lines? {}'.format(
+            OUTPUT_MATCH.keys()
+        )
