@@ -4,11 +4,7 @@
 # See the NOTICE for more information.
 
 import errno
-
-try:
-    import unittest.mock as mock
-except ImportError:
-    import mock
+import unittest.mock as mock
 
 import gunicorn.pidfile
 
