@@ -5,11 +5,7 @@
 
 from contextlib import contextmanager
 import os
-
-try:
-    import unittest.mock as mock
-except ImportError:
-    import mock
+import unittest.mock as mock
 
 import pytest
 
