@@ -5,12 +5,9 @@
 #
 # Example code from Eventlet sources
 
-from wsgiref.validate import validator
-
 from gunicorn import __version__
 
 
-@validator
 def app(environ, start_response):
     """Simplest possible application object"""
 
@@ -24,8 +21,7 @@ def app(environ, start_response):
     response_headers = [
         ('Content-type', 'text/plain'),
         ('Content-Length', str(len(data))),
-        ('X-Gunicorn-Version', __version__),
-        ("Test", "test тест"),
+        ('X-Gunicorn-Version', __version__)
     ]
     start_response(status, response_headers)
     return iter([data])
